+++ packages/gov-ui-kit/CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+-   Updated `HeaderDao` dropdown item styling
+-   Updated `Dropdown` `:hover:focus` styling
+
 ## [0.3.0] - 2023-10-10
 
 ### Changed